requirements.txt
absl-py==2.1.0
altair==5.3.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
asttokens==2.4.1
attrs==23.2.0
autopep8==2.1.0
black==21.4b2
blinker==1.7.0
Brotli==1.1.0
cachetools==5.3.3
Camelot==12.6.29
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
cloudpickle==3.0.0
configobj==5.0.8
configparser==7.0.0
contourpy==1.2.1
cryptography==42.0.7
cycler==0.12.1
decorator==4.4.2
distro==1.9.0
Elixir==0.7.1
etelemetry==0.3.1
exceptiongroup==1.2.0
executing==2.0.1
ffmpeg-python==0.2.0
filelock==3.14.0
fitz==0.0.1.dev2
fonttools==4.51.0
fsspec==2024.3.1
future==1.0.0
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.2
google-api-core==1.34.1
google-api-python-client==2.128.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-vision==1.0.0
google-generativeai==0.5.2
googleapis-common-protos==1.63.0
grpcio==1.63.0
grpcio-status==1.48.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
hydra-core==1.3.2
idna==3.7
imageio==2.34.0
imageio-ffmpeg==0.4.9
importlib-metadata==6.11.0
iopath==0.1.9
ipython==8.24.0
isodate==0.6.1
isort==5.13.2
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
layoutparser==0.3.4
looseversion==1.3.0
lxml==5.2.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
moviepy==1.0.3
mpmath==1.3.0
mutagen==1.47.0
mypy-extensions==1.0.0
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
openai==1.3.4
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.2
parso==0.8.4
pathlib==1.0.1
pathspec==0.12.1
pbr==6.0.0
pdf2image==1.17.0
pdfminer==20191125
pdfminer.six==20231228
pdfplumber==0.11.0
pexpect==4.9.0
Pillow==9.4.0
portalocker==2.8.2
proglog==0.1.10
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==3.20.3
prov==2.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.7
pycodestyle==2.11.1
pycparser==2.22
pycryptodome==3.20.0
pycryptodomex==3.20.0
pydantic==2.7.0
pydantic_core==2.18.1
pydeck==0.8.1b0
pydot==2.0.0
Pygments==2.17.2
PyMuPDF==1.24.4
PyMuPDFb==1.24.3
pyparsing==3.1.2
PyPDF2==3.0.1
pypdfium2==4.30.0
pytesseract==0.3.10
python-dateutil==2.9.0.post0
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
rdflib==7.0.0
referencing==0.34.0
regex==2024.4.16
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
scipy==1.13.0
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==0.7.10
sqlalchemy-migrate==0.11.0
sqlparse==0.5.0
stack-data==0.6.3
streamlit==1.28.2
sympy==1.12
tabulate==0.9.0
Tempita==0.5.2
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==2.4.0
tiktoken==0.5.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.3.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
traits==6.3.2
triton==2.3.0
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.1
validators==0.28.0
watchdog==4.0.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.3
xlrd==0.7.1
xlwt==0.7.2
yacs==0.1.8
yt-dlp==2024.4.9
zipp==3.18.1


